feat(Schema.from_definition) parse @-definitions
